MifosBaseModel organized
