+++ mifosng-android/src/main/java/com/mifos/objects/client/Status.java
@@ -9,11 +9,10 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 import com.mifos.api.local.MifosDatabase;
-import com.mifos.objects.MifosBaseModel;
+import com.mifos.api.local.MifosBaseModel;
 import com.raizlabs.android.dbflow.annotation.Column;
 import com.raizlabs.android.dbflow.annotation.PrimaryKey;
 import com.raizlabs.android.dbflow.annotation.Table;
-import com.raizlabs.android.dbflow.structure.BaseModel;
 
 /**
  * Created by ishankhanna on 09/02/14.
